Some code clean up to be more readable, fixed some DRY principle cases.

+++ header_files/draw_functions.h
@@ -33,7 +33,7 @@
     #include <stdlib.h>
 #endif
 
-const void drawline(Pixel **buffer, float x1, float y1, float x2, float y2, const int red, const int green, const int blue);
+const void drawline(Pixel **buffer, float x1, float y1, float x2, float y2, const float red, const float green, const float blue);
 
 const void filltriangle(Pixel **pixels, float **depth_buffer, Triangle *t, const Global light, const Global camera, const float red, const float green, const float blue);
 const void fillnorthway(Pixel **pixels, float **depth_buffer, const Triangle t, const Global light, const Global camera, const float winding, const Vector obj_color);
